Fix timestamp

+++ platforms/android/aps-sdk/airplay/src/main/cpp/IAirPlayMirrorHandler.cpp
@@ -69,11 +69,11 @@ void IAirPlayMirrorHandler::on_mirror_stream_data(
       jbyteArray byte_array = env->NewByteArray(p->payload_size);
       env->SetByteArrayRegion(byte_array, 0, p->payload_size,
                               (jbyte *)(p->payload));
-      uint32_t timestamp = normalize_ntp_to_ms(p->timestamp);
+      jlong timestamp = (uint32_t)normalize_ntp_to_ms(p->timestamp);
       // convert the value
       __android_log_print(ANDROID_LOG_DEBUG, LOG_TAG,
-                          "================================== Mirror stream data packet timestamp: %d", timestamp);
-      env->CallVoidMethod(jvm_obj_, mid, byte_array, timestamp);
+                          "================================== Mirror stream data packet timestamp: %ld", timestamp);
+      env->CallVoidMethod(jvm_obj_, mid, byte_array, (jlong)timestamp);
       env->DeleteLocalRef(byte_array);
     } else {
       __android_log_write(ANDROID_LOG_ERROR, LOG_TAG,
@@ -187,8 +187,8 @@ void IAirPlayMirrorHandler::on_audio_stream_data(
       env->SetByteArrayRegion(byte_array, 0, payload_length,
                               (jbyte *)(p->payload));
       __android_log_print(ANDROID_LOG_DEBUG, LOG_TAG,
-                          "================================== Audio stream data packet timestamp: %d", p->timestamp);
-      env->CallVoidMethod(jvm_obj_, mid, byte_array, p->timestamp);
+                          "================================== Audio stream data packet timestamp: %ld", p->timestamp);
+      env->CallVoidMethod(jvm_obj_, mid, byte_array, (jlong)(p->timestamp));
       env->DeleteLocalRef(byte_array);
     } else {
       __android_log_write(ANDROID_LOG_ERROR, LOG_TAG,
